+++ pandas/tests/scalar/timedelta/test_timedelta.py
@@ -4,56 +4,14 @@ from datetime import timedelta
 import numpy as np
 import pytest
 
-from pandas._libs.tslibs import NaT, Timestamp, iNaT
+from pandas._libs.tslibs import NaT, iNaT
 
 import pandas as pd
 from pandas import Timedelta, TimedeltaIndex, offsets, to_timedelta
 import pandas._testing as tm
 
 
-class TestTimedeltaArithmetic:
-    def test_arithmetic_overflow(self):
-        with pytest.raises(OverflowError):
-            Timestamp("1700-01-01") + Timedelta(13 * 19999, unit="D")
-
-        with pytest.raises(OverflowError):
-            Timestamp("1700-01-01") + timedelta(days=13 * 19999)
-
-    def test_array_timedelta_floordiv(self):
-        # deprecated GH#19761, enforced GH#29797
-        ints = np.array([1349654400, 1349740800, 1349827200, 1349913600]) * 10 ** 9
-
-        with pytest.raises(TypeError, match="Invalid dtype"):
-            ints // Timedelta(1, unit="s")
-
-    def test_ops_error_str(self):
-        # GH 13624
-        td = Timedelta("1 day")
-
-        for left, right in [(td, "a"), ("a", td)]:
-
-            with pytest.raises(TypeError):
-                left + right
-
-            with pytest.raises(TypeError):
-                left > right
-
-            assert not left == right
-            assert left != right
-
-    def test_ops_notimplemented(self):
-        class Other:
-            pass
-
-        other = Other()
-
-        td = Timedelta("1 day")
-        assert td.__add__(other) is NotImplemented
-        assert td.__sub__(other) is NotImplemented
-        assert td.__truediv__(other) is NotImplemented
-        assert td.__mul__(other) is NotImplemented
-        assert td.__floordiv__(other) is NotImplemented
-
+class TestTimedeltaUnaryOps:
     def test_unary_ops(self):
         td = Timedelta(10, unit="d")
 
@@ -66,102 +24,6 @@ class TestTimedeltaArithmetic:
         assert abs(td) == td
         assert abs(-td) == td
         assert abs(-td) == Timedelta("10d")
-
-
-class TestTimedeltaComparison:
-    def test_compare_tick(self, tick_classes):
-        cls = tick_classes
-
-        off = cls(4)
-        td = off.delta
-        assert isinstance(td, Timedelta)
-
-        assert td == off
-        assert not td != off
-        assert td <= off
-        assert td >= off
-        assert not td < off
-        assert not td > off
-
-        assert not td == 2 * off
-        assert td != 2 * off
-        assert td <= 2 * off
-        assert td < 2 * off
-        assert not td >= 2 * off
-        assert not td > 2 * off
-
-    def test_comparison_object_array(self):
-        # analogous to GH#15183
-        td = Timedelta("2 days")
-        other = Timedelta("3 hours")
-
-        arr = np.array([other, td], dtype=object)
-        res = arr == td
-        expected = np.array([False, True], dtype=bool)
-        assert (res == expected).all()
-
-        # 2D case
-        arr = np.array([[other, td], [td, other]], dtype=object)
-        res = arr != td
-        expected = np.array([[True, False], [False, True]], dtype=bool)
-        assert res.shape == expected.shape
-        assert (res == expected).all()
-
-    def test_compare_timedelta_ndarray(self):
-        # GH11835
-        periods = [Timedelta("0 days 01:00:00"), Timedelta("0 days 01:00:00")]
-        arr = np.array(periods)
-        result = arr[0] > arr
-        expected = np.array([False, False])
-        tm.assert_numpy_array_equal(result, expected)
-
-    @pytest.mark.skip(reason="GH#20829 is reverted until after 0.24.0")
-    def test_compare_custom_object(self):
-        """
-        Make sure non supported operations on Timedelta returns NonImplemented
-        and yields to other operand (GH#20829).
-        """
-
-        class CustomClass:
-            def __init__(self, cmp_result=None):
-                self.cmp_result = cmp_result
-
-            def generic_result(self):
-                if self.cmp_result is None:
-                    return NotImplemented
-                else:
-                    return self.cmp_result
-
-            def __eq__(self, other):
-                return self.generic_result()
-
-            def __gt__(self, other):
-                return self.generic_result()
-
-        t = Timedelta("1s")
-
-        assert not (t == "string")
-        assert not (t == 1)
-        assert not (t == CustomClass())
-        assert not (t == CustomClass(cmp_result=False))
-
-        assert t < CustomClass(cmp_result=True)
-        assert not (t < CustomClass(cmp_result=False))
-
-        assert t == CustomClass(cmp_result=True)
-
-    @pytest.mark.parametrize("val", ["string", 1])
-    def test_compare_unknown_type(self, val):
-        # GH20829
-        t = Timedelta("1s")
-        with pytest.raises(TypeError):
-            t >= val
-        with pytest.raises(TypeError):
-            t > val
-        with pytest.raises(TypeError):
-            t <= val
-        with pytest.raises(TypeError):
-            t < val
 
 
 class TestTimedeltas:
